backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
rembg==2.0.50
requests==2.31.0
python-dotenv==1.0.0
torch==2.0.1
torchvision==0.15.2
transformers==4.35.0
accelerate==0.24.1
diffusers==0.21.4
safetensors==0.4.0
